Code cleanup. Isolated text highlight to function
