+++ app/build.gradle
@@ -17,10 +17,6 @@ android {
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
-    dexOptions {
-        preDexLibraries = false
-    }
-
     buildTypes {
         release {
             minifyEnabled false
