+++ app/src/main/java/com/lukasz/galinski/fluffy/view/account/RegisterScreen.kt
@@ -1,19 +1,16 @@
 package com.lukasz.galinski.fluffy.view.account
 
 import android.os.Bundle
-import android.text.Spannable
-import android.text.SpannableString
-import android.text.style.ForegroundColorSpan
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
-import com.lukasz.galinski.fluffy.R
+import com.lukasz.galinski.fluffy.common.highlightSelectedTextRange
 import com.lukasz.galinski.fluffy.databinding.RegisterScreenFragmentBinding
 
 private const val HIGHLIGHTED_TERMS_SPANS_COUNT = 35
 private const val HIGHLIGHTED_LOGIN_SPANS_COUNT = 5
+private const val HIGHLIGHTED_COLOR = "#7F3DFF"
 
 class RegisterScreen : Fragment() {
 
@@ -30,37 +27,23 @@ class RegisterScreen : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        highlightTextParts()
+    }
+
+    private fun highlightTextParts(){
         val checkBoxText = registerBinding.termsCheckbox
         val loginLabel = registerBinding.existingAccountInfo
         checkBoxText.text = highlightSelectedTextRange(
             checkBoxText.text,
             checkBoxText.text.length - HIGHLIGHTED_TERMS_SPANS_COUNT,
-            checkBoxText.text.length
+            checkBoxText.text.length,
+            HIGHLIGHTED_COLOR
         )
         loginLabel.text = highlightSelectedTextRange(
             loginLabel.text,
             loginLabel.text.length - HIGHLIGHTED_LOGIN_SPANS_COUNT,
-            loginLabel.text.length
+            loginLabel.text.length,
+            HIGHLIGHTED_COLOR
         )
-    }
-
-    private fun highlightSelectedTextRange(
-        message: CharSequence,
-        startPosition: Int,
-        endPosition: Int
-    ): SpannableString {
-        val spannable = SpannableString(message)
-        spannable.setSpan(
-            ForegroundColorSpan(
-                ContextCompat.getColor(
-                    activity?.applicationContext!!,
-                    R.color.purple_primary
-                )
-            ),
-            startPosition, // start
-            endPosition, // end
-            Spannable.SPAN_EXCLUSIVE_INCLUSIVE
-        )
-        return spannable
     }
 }
